QueryProvider update with Hydrate boundry
