+++ b-fit/lib/getQueryClient.tsx
@@ -1,0 +1,27 @@
+import { QueryClient, isServer } from "@tanstack/react-query";
+
+function makeQueryClient() {
+  return new QueryClient({
+    defaultOptions: {
+      queries: {
+        staleTime: 60 * 1000,
+      },
+    },
+  });
+}
+
+let browserQueryClient: QueryClient | undefined = undefined;
+
+/**
+ * Function to return query client based on where it is used.
+ * Server: Return a new client query each time to prevent duplicate data based on queries
+ * Client: use existing client to make sure data is not lost
+ */
+export function getQueryClient() {
+  if (isServer) {
+    return makeQueryClient();
+  } else {
+    if (!browserQueryClient) browserQueryClient = makeQueryClient();
+    return browserQueryClient;
+  }
+}
